requirements.txt
streamlit>=1.18
openai>=1.0.0
chromadb>=0.4.0
tiktoken
requests
beautifulsoup4
fpdf==1.7.2
python-Levenshtein
numpy
python-dotenv
